separate generic DB helpers into db_p.h
when the QPDB is implemented, we will need to have both qpdb_p.h and
rbtdb_p.h. to prevent name collisions or code duplication, we here
create a private header file for generic database implementations, for
structures and macros that will be used by both databases.

we also rename some functions and structs to more specifically refer to
the RBT database, to avoid namespace collision with similar things that
will be needed by the QP database later.