Check that a zone that serves AAAA is served over IPv6
named-checkzone will now, as part of the zone's integrity checks,
look to see if there are AAAA records being served and if so check
that the nameservers have AAAA records.  This is a sometime times
over looked check that, if not met, can mean that a service that is
supposed to run over IPv6 will not when there is no IPv4 reachability.