Remove the contention when pruning RBTDB nodes
In #4383, we reported:

    In some older BIND 9 branches, the extra queuing overhead eliminated by
    this change could be remotely exploited to cause excessive memory use.
    Due to architectural shift, this branch is not vulnerable to that issue,
    but applying the fix to the latter is nevertheless deemed prudent for
    consistency and to make the code future-proof.

The single queue for pruning the nodes increased the lock contention to
the level where cleaning up nodes from the RBTDB would take too long and
the memory that the cache was using would grow indefinitely.

This commit reverts the change to the pruning mechanism as the 9.18+
release was not affected by affor mentioned problem with queueing
overhead.