Replace the two dns_fixedname_t in ns_query_t with hashes
The ns_query_t structure has a mechanism to check whether we are not in
a recursion loop.  Previously, the full qname and qdomain was stored in
the structure as dns_fixedname_t which blew up the size of the
ns_query_t by >= 512 bytes.  Replace the two fixed names with full
hashes of the qname and qdomain.  This reduces the size of the recparam
structure to roughly less than 20 bytes.