Allow several rcu_barrier() calls when waiting for memory reclamation
There is a possiblity that the call_rcu() callback for which the
rcu_barrier() was waiting for doesn't free up any resources, in
which case the code breaks from the loop and could leave some
nested callbacks still running.

Allow a sane amount of such nested callbacks before giving up.